dev-requirements.txt
-i https://pypi.org/simple
astroid==3.0.3; python_full_version >= '3.8.0'
autopep8==2.0.4; python_version >= '3.6'
dill==0.3.8; python_version >= '3.11'
isort==5.13.2; python_full_version >= '3.8.0'
mccabe==0.7.0; python_version >= '3.6'
pep8==1.7.1
platformdirs==4.2.0; python_version >= '3.8'
pycodestyle==2.11.1; python_version >= '3.8'
pylint==3.0.3; python_full_version >= '3.8.0'
pylint-django==2.5.5; python_version >= '3.7' and python_version < '4.0'
pylint-plugin-utils==0.8.2; python_version >= '3.7' and python_version < '4.0'
tomlkit==0.12.3; python_version >= '3.7'
